vulns: extract Identify from Scanner
Move repo_url, tag, FORGES and TAG_PATTERNS into a new
Homebrew::Vulns::Identify module so the upcoming advisory-match
dev-cmd can share the URL-parsing helpers without pulling in Scanner.
Scanner requires the new module and calls it directly. No behaviour
change; specs moved to identify_spec.rb.
